16. In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A+

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,7 +6,7 @@
  */
 int main(void)
 {
-	unsigned long int a = 1, b = 2, c,  count = 0;
+	unsigned long long int a = 1, b = 2, c,  count = 0;
 
 	printf("%ld, ", a);
 	count++;
